Refs #34191 - rename actions to YumInstallablePackages
