+++ webpack/components/extensions/HostDetails/Tabs/RemoteExecutionActions.js
@@ -5,7 +5,7 @@ import { getResponseErrorMsgs } from '../../../../utils/helpers';
 import { renderTaskStartedToast } from '../../../../scenes/Tasks/helpers';
 import { ERRATA_SEARCH_QUERY } from './ErrataTab/HostErrataConstants';
 import { TRACES_SEARCH_QUERY } from './TracesTab/HostTracesConstants';
-import { PACKAGE_SEARCH_QUERY } from '../ApplicablePackages/ApplicablePackagesConstants';
+import { PACKAGE_SEARCH_QUERY } from '../YumInstallablePackages/YumInstallablePackagesConstants';
 
 const errorToast = (error) => {
   const message = getResponseErrorMsgs(error.response);
