+++ webpack/components/extensions/HostDetails/YumInstallablePackages/YumInstallablePackagesConstants.js
@@ -1,0 +1,3 @@
+export const HOST_YUM_INSTALLABLE_PACKAGES_KEY = 'HOST_YUM_INSTALLABLE_PACKAGES';
+export const PACKAGE_SEARCH_QUERY = 'Package search query';
+export default HOST_YUM_INSTALLABLE_PACKAGES_KEY;
